+++ truck-assembly/src/lib.rs
@@ -1,1 +1,18 @@
-pub mod graph;
+//! Crate for handling assembly structures
+
+#![deny(clippy::all, rust_2018_idioms)]
+#![warn(
+    missing_docs,
+    missing_debug_implementations,
+    trivial_casts,
+    trivial_numeric_casts,
+    unsafe_code,
+    unstable_features,
+    unused_import_braces,
+    unused_qualifications
+)]
+
+/// Abstract DAG structure
+pub mod dag;
+/// Assembly structure
+pub mod assy;